fix: update virtualenv building so llamastack- prefix is not added, make notebook experience easier
